Rename name of the page without dashes.
